% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsexample-package.R
\docType{package}
\name{vsexample-package}
\alias{vsexample}
\alias{vsexample-package}
\title{vsexample: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/vsexample}
  \item Report bugs at \url{https://github.com/DavisVaughan/vsexample/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
